better exports, start on new trivia
