+++ src/core/commands/index.ts
@@ -1,2 +1,4 @@
 export { default as Command } from './command';
 export { default as registry } from './registry';
+export { default as CommandRunner } from './runner';
+export * from './types';
