+++ src/core/model/index.ts
@@ -1,0 +1,5 @@
+export { default as BotMessage, SavedBotMessage } from './bot-message';
+export { default as Channel } from './channel';
+export { default as Message, SavedMessage } from './message';
+export { ID } from './types';
+export { default as User } from './user';
