+++ src/core/oop/decorators/index.ts
@@ -1,1 +1,3 @@
+export * from './arg';
 export * from './cmd';
+export * from './validation';
