+++ src/core/oop/index.ts
@@ -1,3 +1,4 @@
 export { Controller } from './controller';
 export { register } from './register';
 export * from './store';
+export * from './decorators';
