+++ src/core/oop/types.ts
@@ -1,0 +1,7 @@
+export type Validator = (
+    arg: string,
+    index: number,
+    ctx: { args: string[] }
+) => true | string | Promise<true | string>;
+
+export type InjectableValidator = (...args: any[]) => Validator;
