+++ src/modules/trivia-legacy/fetch.ts
