+++ src/modules/trivia/trivia.store.ts
@@ -1,0 +1,31 @@
+import { Store, Channel } from 'core';
+
+interface ITriviaStore {
+    enabledCategories: number[];
+    noReply: boolean;
+    autostart: boolean;
+
+    game: {
+        question: string;
+        options: string[];
+        answer: string;
+        difficulty: string;
+    } | null;
+}
+
+interface TriviaStore extends ITriviaStore {}
+
+class TriviaStore extends Store<ITriviaStore> {
+    initial = {
+        enabledCategories: [],
+        noReply: true,
+        autostart: false,
+        game: null,
+    };
+
+    constructor(channel: Channel) {
+        super(`trivia:${channel.id}`);
+    }
+}
+
+export default TriviaStore;
